rename: 파일&변수명 수정 manufactureFilter -> ProductsFilter
